requirements.txt
appnope==0.1.2
backcall==0.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
decorator==5.1.0
google-api-core==2.3.0
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-pubsub==2.9.0
googleapis-common-protos==1.54.0
grpc-google-iam-v1==0.12.3
grpcio==1.42.0
grpcio-status==1.42.0
httplib2==0.20.2
idna==3.3
ipython==7.30.1
jedi==0.18.1
kafka-python==2.0.2
libcst==0.3.23
matplotlib-inline==0.1.3
mypy==0.920
mypy-extensions==0.4.3
numpy==1.21.4
pandas==1.3.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.23
proto-plus==1.19.8
protobuf==3.19.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
rsa==4.8
six==1.16.0
tomli==2.0.0
torch==1.10.0
torchvision==0.10.0
traitlets==5.1.1
typing-inspect==0.7.1
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.2.5
